move ConventionService to services_conventions.py
